+++ lib/uuidv7.ex
@@ -1,12 +1,10 @@
 defmodule UUIDv7 do
   @moduledoc """
-  A UUID v7 implementation and `Ecto.Type` for Elixir - based on Rust.
+  A UUID v7 implementation and `Ecto.Type` for Elixir.
 
-  This library defers the UUID v7 implementation to the Rust create [UUID](https://crates.io/crates/uuid)
-  using an Erlang NIF. It includes an `Ecto.Type` to (auto-)generate version 7 UUIDs in `Ecto.Schema` and beyond.
+  The RFC for the version 7 UUID: [RFC 9562](https://datatracker.ietf.org/doc/rfc9562/).
 
-  Thanks to Rust, it is ~72% faster in generating version 7 UUIDs than the Elixir implementation
-  of version 4 UUIDs by Ecto. See the benchmarks for more details.
+  This library includes an `Ecto.Type` to (auto-)generate version 7 UUIDs in `Ecto.Schema` and beyond.
 
   ## Installation
 
@@ -14,7 +12,7 @@ defmodule UUIDv7 do
 
   ```elixir
   def deps do
-    [{:uuidv7, "~> 0.2"}]
+    [{:uuidv7, "~> 1.0"}]
   end
   ```
 
@@ -23,99 +21,75 @@ defmodule UUIDv7 do
   In your database schema, change primary key attribute from `:binary_id` to `UUIDv7`:
 
   ```elixir
-  def App.Schemas.User do
+  def MyApp.Schemas.User do
     @primary_key {:id, UUIDv7, autogenerate: true}
   end
   ```
-
-  ## Benchmark
-
-  Run `mix deps.get` and then `mix run scripts/bench.exs` to run the benchmark on your computer.
-
-  ```zsh
-  Operating System: macOS
-  CPU Information: Apple M2 Pro
-  Number of Available Cores: 10
-  Available memory: 16 GB
-  Elixir 1.14.2
-  Erlang 25.1.2
-
-  Benchmark suite executing with the following configuration:
-  warmup: 5 s
-  time: 10 s
-  memory time: 5 s
-  reduction time: 0 ns
-  parallel: 1
-  inputs: none specified
-  Estimated total run time: 1 min
-
-  Benchmarking ecto (uuid v4) ...
-  Benchmarking uniq (uuid v7) ...
-  Benchmarking uuidv7 ...
-
-  Name                     ips        average  deviation         median         99th %
-  uuidv7                1.75 M      570.22 ns  ±3940.19%         500 ns         667 ns
-  uniq (uuid v7)        1.07 M      937.20 ns  ±1852.78%         916 ns        1000 ns
-  ecto (uuid v4)        1.02 M      978.17 ns  ±1593.54%         958 ns        1042 ns
-
-  Comparison:
-  uuidv7                1.75 M
-  uniq (uuid v7)        1.07 M - 1.64x slower +366.98 ns
-  ecto (uuid v4)        1.02 M - 1.72x slower +407.95 ns
-
-  Memory usage statistics:
-
-  Name                   average  deviation         median         99th %
-  uuidv7                   104 B     ±0.00%          104 B          104 B
-  uniq (uuid v7)        214.00 B     ±2.47%          216 B          216 B
-  ecto (uuid v4)        214.00 B     ±2.47%          216 B          216 B
-
-  Comparison:
-  uuidv7                   104 B
-  uniq (uuid v7)        214.00 B - 2.06x memory usage +110.00 B
-  ecto (uuid v4)        214.00 B - 2.06x memory usage +110.00 B
-  ```
-
-  ## Credits
-
-  This library is based on the Rust library [UUID](https://crates.io/crates/uuid). Thank you!
-
-  The `Ecto.Type` is heavily borrowed from from [Ecto](https://github.com/elixir-ecto/ecto). Thanks!
   """
 
-  @version Mix.Project.config()[:version]
+  use Ecto.Type
 
-  use RustlerPrecompiled,
-    otp_app: :uuidv7,
-    crate: "uuidv7",
-    base_url: "https://github.com/martinthenth/uuidv7/releases/download/v#{@version}",
-    nif_versions: ["2.16", "2.17"],
-    targets:
-      Enum.uniq(["aarch64-unknown-linux-musl" | RustlerPrecompiled.Config.default_targets()]),
-    force_build: System.get_env("FORCE_BUILD") in ["1", "true"],
-    version: @version
+  @typedoc """
+  A hex-encoded UUID string.
+  """
+  @type t :: <<_::288>>
 
-  use UUIDv7.Type
+  @typedoc """
+  A raw binary representation of a UUID.
+  """
+  @type raw :: <<_::128>>
+
+  @doc false
+  @spec type() :: :uuid
+  def type, do: :uuid
+
+  defdelegate cast(value), to: Ecto.UUID
+
+  defdelegate cast!(value), to: Ecto.UUID
+
+  defdelegate dump(value), to: Ecto.UUID
+
+  defdelegate dump!(value), to: Ecto.UUID
+
+  defdelegate load(value), to: Ecto.UUID
+
+  defdelegate load!(value), to: Ecto.UUID
+
+  @doc false
+  @spec autogenerate() :: t()
+  def autogenerate, do: generate()
 
   @doc """
   Generates a version 7 UUID.
   """
   @spec generate() :: t()
-  def generate, do: :erlang.nif_error(:nif_not_loaded)
+  def generate, do: Ecto.UUID.cast!(bingenerate())
 
   @doc """
   Generates a version 7 UUID based on the timestamp (ms).
   """
   @spec generate(non_neg_integer()) :: t()
-  def generate(ms), do: generate_from_ms(ms)
+  def generate(milliseconds), do: milliseconds |> bingenerate() |> Ecto.UUID.cast!()
 
   @doc """
-  Generates a version 7 UUID based on the timestamp (ms).
+  Generates a version 7 UUID in the binary format.
   """
-  @spec generate_from_ms(non_neg_integer()) :: t()
-  def generate_from_ms(_ms), do: :erlang.nif_error(:nif_not_loaded)
+  @spec bingenerate() :: raw()
+  def bingenerate, do: :millisecond |> System.system_time() |> bingenerate()
 
-  @doc false
-  @spec autogenerate() :: t()
-  def autogenerate, do: generate()
+  @doc """
+  Generates a version 7 UUID in the binary format based on the timestamp (ms).
+  """
+  @spec bingenerate(non_neg_integer()) :: raw()
+  def bingenerate(milliseconds) do
+    <<u1::12, u2::62, _::6>> = :crypto.strong_rand_bytes(10)
+    <<milliseconds::48, 7::4, u1::12, 2::2, u2::62>>
+  end
+
+  @doc """
+  Extracts the timestamp (ms) from the version 7 UUID.
+  """
+  @spec timestamp(t() | raw()) :: non_neg_integer()
+  def timestamp(<<milliseconds::48, 7::4, _::76>>), do: milliseconds
+  def timestamp(<<_::288>> = uuid), do: uuid |> Ecto.UUID.dump!() |> timestamp()
 end