+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 1.0.0
+
+- Removes the dependency on Rust, it is now an Elixir-based implementation
+  - Adds a test suite to ensure the implementation remains compliant to the RFC
+  - RFC: [](https://datatracker.ietf.org/doc/rfc9562/)
+- Renames `generate_from_ms/1` to `generate/1`
+- Adds `timestamp/1` to get the timestamp (milliseconds) from a version 7 UUID.
+- Graduates the library to version `1.0.0` 🎉
+
 # 0.2.1
 
 - Add more build targets for the NIFs
@@ -5,7 +14,7 @@
 
 # 0.2.0
 
-- Renames `generate_from_ns/1` to `generate_from_ms/2`
+- Renames `generate_from_ns/1` to `generate_from_ms/1`
 - Refactors the `Ecto.Type` code into it's own module as a macro
 
 # 0.1.0